implement useDataPersistence
