+++ src/App.vue
@@ -92,7 +92,7 @@
       :unions="unions"
       :is-editing="!!editingMember"
       @close="closeModal"
-      @save="saveMember"
+      @save="handleSaveMember"
     />
 
     <!-- Modal for Add/Edit Union -->
@@ -103,7 +103,7 @@
       :unions="unions"
       :is-editing="!!editingUnion"
       @close="closeUnionModal"
-      @save="saveUnion"
+      @save="handleSaveUnion"
     />
 
     <!-- Delete Confirmation Modal -->
@@ -259,7 +259,7 @@
           <SettingsForm
             :members="members"
             :current-settings="settings"
-            @save="saveSettings"
+            @save="handleSaveSettings"
             @cancel="closeSettingsModal"
           />
         </div>
@@ -290,6 +290,7 @@ import {
 } from '@/types'
 import { useFileOperations } from '@/composables/useFileOperations'
 import { useMemberUtils } from '@/composables/useMemberUtils'
+import { useDataPersistence } from '@/composables/useDataPersistence'
 import Header from '@/components/Header.vue'
 import FamilyTree from '@/components/FamilyTree.vue'
 import MemberDetail from '@/components/MemberDetail.vue'
@@ -301,8 +302,6 @@ import AnniversaryDrawer from '@/components/AnniversaryDrawer.vue'
 import AnniversaryToggleButton from '@/components/AnniversaryToggleButton.vue'
 
 // State
-const members = ref<Member[]>([])
-const unions = ref<Union[]>([])
 const selectedMember = ref<Member | null>(null)
 const showAddForm = ref(false)
 const editingMember = ref<Member | null>(null)
@@ -326,43 +325,48 @@ const settings = ref({
 // Anniversary drawer state
 const showAnniversaryDrawer = ref(false)
 
-// File operations composable
+// Composables
 const { fileInput, exportToJson, importFromJson, handleFileImport } =
   useFileOperations()
-
-// Member utils composable
 const { getFullName } = useMemberUtils()
+const {
+  members,
+  unions,
+  getNextId,
+  saveToLocalStorage,
+  loadSettings,
+  saveSettings,
+  clearAllData,
+  saveMember,
+  deleteMember,
+  saveUnion,
+  initializeData,
+  getMemberById,
+  getUnionById,
+  getAllMembers,
+  getAllUnions,
+  getFamilyData,
+  setFamilyData,
+} = useDataPersistence()
 
 // Load data from localStorage or default
 onMounted(() => {
-  const savedData = localStorage.getItem('family-data')
-  if (savedData) {
-    const parsedData = JSON.parse(savedData)
-    // Migrate existing data to ensure proper structure
-    const migratedData = migrateData(parsedData)
-    members.value = migratedData.members
-    unions.value = migratedData.unions
-  }
+  // Initialize data from persistence layer
+  initializeData()
 
   // Load settings
-  const savedSettings = localStorage.getItem('family-settings')
-  if (savedSettings) {
-    const parsedSettings = JSON.parse(savedSettings)
-    settings.value = {
-      ...settings.value,
-      ...parsedSettings,
-      // Ensure anniversary settings have defaults
-      showBirthdays: parsedSettings.showBirthdays ?? true,
-      showMarriages: parsedSettings.showMarriages ?? false,
-      showDeaths: parsedSettings.showDeaths ?? false,
-    }
+  const defaultSettings = {
+    defaultMemberId: null as number | null,
+    showDates: true,
+    showBirthdays: true,
+    showMarriages: false,
+    showDeaths: false,
   }
+  settings.value = loadSettings(defaultSettings)
 
   // Set default selected member if specified
   if (settings.value.defaultMemberId) {
-    const defaultMember = members.value.find(
-      (m) => m.id === settings.value.defaultMemberId,
-    )
+    const defaultMember = getMemberById(settings.value.defaultMemberId)
     if (defaultMember) {
       // Use nextTick to ensure the FamilyTree component is mounted
       nextTick(() => {
@@ -375,110 +379,7 @@ onMounted(() => {
   }
 })
 
-// Migration function to ensure proper data structure
-const migrateData = (data: any): FamilyData => {
-  let members: any[] = []
-  let unions: Union[] = []
-
-  // Handle old format where unions were embedded in members
-  if (Array.isArray(data)) {
-    // Old format: array of members with embedded unions
-    members = data.map((member) => ({
-      ...member,
-      unions: member.unions || [],
-      spouseIds: member.spouseIds || [], // Keep for backward compatibility
-    }))
-
-    // Extract unions from members
-    members.forEach((member) => {
-      if (member.unions && Array.isArray(member.unions)) {
-        member.unions.forEach((union: any) => {
-          // Avoid duplicates
-          if (!unions.find((u) => u.id === union.id)) {
-            unions.push(union)
-          }
-        })
-      }
-    })
-  } else if (data.members && data.unions) {
-    // New format: separate members and unions
-    members = data.members.map((member: any) => ({
-      ...member,
-      unions: member.unions || [],
-      spouseIds: member.spouseIds || [],
-    }))
-    unions = data.unions || []
-  } else {
-    // Fallback: empty data
-    return { members: [], unions: [] }
-  }
-
-  // Migrate parent1Id/parent2Id to parentUnionId
-  members = members.map((member) => {
-    const migratedMember = { ...member }
-
-    // Remove old parent properties
-    delete migratedMember.parent1Id
-    delete migratedMember.parent2Id
-
-    // Ensure parentUnionId exists (will be undefined if no parents)
-    if (!migratedMember.hasOwnProperty('parentUnionId')) {
-      migratedMember.parentUnionId = undefined
-    }
-
-    // Ensure middleNames is an array
-    if (!Array.isArray(migratedMember.middleNames)) {
-      migratedMember.middleNames = []
-    }
-
-    return migratedMember
-  })
-
-  // Ensure all unions have childrenIds array
-  unions = unions.map((union) => ({
-    ...union,
-    childrenIds: Array.isArray(union.childrenIds) ? union.childrenIds : [],
-  }))
-
-  // Build parentUnionId dynamically from unions
-  members = buildParentUnionIds(members, unions)
-
-  return { members, unions }
-}
-
-// Function to build parentUnionId dynamically from unions
-const buildParentUnionIds = (members: Member[], unions: Union[]): Member[] => {
-  return members.map((member) => {
-    // Find which union has this member as a child
-    const parentUnion = unions.find(
-      (union) =>
-        union.childrenIds &&
-        Array.isArray(union.childrenIds) &&
-        union.childrenIds.includes(member.id),
-    )
-
-    return {
-      ...member,
-      parentUnionId: parentUnion?.id || undefined,
-    }
-  })
-}
-
-// Computed
-const nextId = computed(() => {
-  return members.value.length > 0
-    ? Math.max(...members.value.map((m) => m.id)) + 1
-    : 1
-})
-
 // Methods
-const saveToLocalStorage = () => {
-  const familyData: FamilyData = {
-    members: members.value,
-    unions: unions.value,
-  }
-  localStorage.setItem('family-data', JSON.stringify(familyData))
-}
 
 const selectMember = (member: Member) => {
   selectedMember.value = member
@@ -510,45 +411,13 @@ const cancelDelete = () => {
   memberToDelete.value = null
 }
 
-const deleteMember = () => {
+const handleDeleteMember = () => {
   if (!memberToDelete.value) return
 
   const memberId = memberToDelete.value.id
 
-  // Remove from members array
-  members.value = members.value.filter((m) => m.id !== memberId)
-
-  // Remove unions that involve the deleted member
-  unions.value = unions.value.filter(
-    (union) => union.member1Id !== memberId && union.member2Id !== memberId,
-  )
-
-  // Update relationships - remove member from parent unions
-  members.value.forEach((member) => {
-    if (member.parentUnionId) {
-      const parentUnion = unions.value.find(
-        (u) => u.id === member.parentUnionId,
-      )
-      if (
-        parentUnion &&
-        parentUnion.childrenIds &&
-        Array.isArray(parentUnion.childrenIds)
-      ) {
-        parentUnion.childrenIds = parentUnion.childrenIds.filter(
-          (id) => id !== memberId,
-        )
-      }
-    }
-  })
-
-  // Remove the member from any union's children lists
-  unions.value.forEach((union) => {
-    if (union.childrenIds && Array.isArray(union.childrenIds)) {
-      union.childrenIds = union.childrenIds.filter(
-        (childId) => childId !== memberId,
-      )
-    }
-  })
+  // Use the composable to delete the member
+  deleteMember(memberId)
 
   // Clear selection if deleted member was selected
   if (selectedMember.value?.id === memberId) {
@@ -579,85 +448,19 @@ const addUnion = () => {
   showUnionForm.value = true
 }
 
-const saveUnion = (formData: UnionFormData) => {
+const handleSaveUnion = (formData: UnionFormData) => {
   console.log('saveUnion called with:', formData)
-  const now = new Date().toISOString()
-  const unionId =
-    editingUnion.value?.id && editingUnion.value.id > 0
-      ? editingUnion.value.id
-      : nextId.value
-
-  // Create or update the union
-  const union: Union = {
-    id: unionId,
-    member1Id: Number(formData.member1Id),
-    member2Id: Number(formData.member2Id),
-    marriageDate: formData.marriageDate,
-    marriagePlace: formData.marriagePlace,
-    divorceDate: formData.divorceDate,
-    divorcePlace: formData.divorcePlace,
-    childrenIds: formData.childrenIds.map(Number),
-  }
-
-  console.log('Created union:', union)
 
   // Check if we're editing an existing union (has real ID) or creating new one
   const isEditingExisting = editingUnion.value && editingUnion.value.id > 0
 
-  if (isEditingExisting) {
-    // Update existing union
-    const unionIndex = unions.value.findIndex((u) => u.id === unionId)
-    if (unionIndex !== -1) {
-      // Handle children relationship changes for existing union
-      const oldUnion = unions.value[unionIndex]
-      const oldChildrenIds = new Set(oldUnion.childrenIds || [])
-      const newChildrenIds = new Set(union.childrenIds || [])
-
-      // Remove children from old parent union
-      const removedChildren = [...oldChildrenIds].filter(
-        (id) => !newChildrenIds.has(id),
-      )
-      removedChildren.forEach((childId) => {
-        const child = members.value.find((m) => m.id === childId)
-        if (child && child.parentUnionId === oldUnion.id) {
-          child.parentUnionId = undefined
-        }
-      })
-
-      // Add children to new parent union
-      const addedChildren = [...newChildrenIds].filter(
-        (id) => !oldChildrenIds.has(id),
-      )
-      addedChildren.forEach((childId) => {
-        const child = members.value.find((m) => m.id === childId)
-        if (child) {
-          child.parentUnionId = union.id
-        }
-      })
-
-      unions.value[unionIndex] = union
-    }
-  } else {
-    // Add new union and set up parent-child relationships
-    if (union.childrenIds && Array.isArray(union.childrenIds)) {
-      union.childrenIds.forEach((childId) => {
-        const child = members.value.find((m) => m.id === childId)
-        if (child) {
-          child.parentUnionId = union.id
-        }
-      })
-    }
-
-    unions.value.push(union)
-  }
+  // Use the composable to save the union
+  const union = saveUnion(formData, isEditingExisting)
 
   console.log('Updated unions array:', {
     totalUnions: unions.value.length,
     isEditingExisting,
   })
-
-  // Rebuild parentUnionIds to ensure consistency
-  members.value = buildParentUnionIds(members.value, unions.value)
 
   saveToLocalStorage()
   closeUnionModal()
@@ -668,93 +471,17 @@ const closeUnionModal = () => {
   editingUnion.value = null
 }
 
-const saveMember = (formData: MemberFormData) => {
-  const now = new Date().toISOString()
+const handleSaveMember = (formData: MemberFormData) => {
+  // Check if we're editing an existing member
+  const isEditing = !!editingMember.value
 
-  // --- Data Sanitization ---
-  // Ensure all incoming IDs from the form are numbers, not strings
-  const sanitizedFormData = {
-    ...formData,
-    parentUnionId: formData.parentUnionId
-      ? Number(formData.parentUnionId)
-      : undefined,
+  // Use the composable to save the member
+  const member = saveMember(formData, isEditing)
+
+  // Set as selected member if it's a new member
+  if (!isEditing) {
+    selectedMember.value = member
   }
-
-  const memberId = editingMember.value?.id || nextId.value
-
-  // --- Handle Parent-Child relationship changes ---
-  const oldParentUnionId = editingMember.value?.parentUnionId
-  const newParentUnionId = sanitizedFormData.parentUnionId
-
-  // If parent union changed, we need to update the union's children list
-  if (oldParentUnionId !== newParentUnionId) {
-    // Remove from old parent union
-    if (oldParentUnionId) {
-      const oldUnion = unions.value.find((u) => u.id === oldParentUnionId)
-      if (
-        oldUnion &&
-        oldUnion.childrenIds &&
-        Array.isArray(oldUnion.childrenIds)
-      ) {
-        oldUnion.childrenIds = oldUnion.childrenIds.filter(
-          (id) => id !== memberId,
-        )
-      }
-    }
-
-    // Add to new parent union
-    if (newParentUnionId) {
-      const newUnion = unions.value.find((u) => u.id === newParentUnionId)
-      if (newUnion) {
-        if (!newUnion.childrenIds) {
-          newUnion.childrenIds = []
-        }
-        if (!newUnion.childrenIds.includes(memberId)) {
-          newUnion.childrenIds.push(memberId)
-        }
-      }
-    }
-  }
-
-  // --- Upsert the member's own data ---
-  if (editingMember.value) {
-    const index = members.value.findIndex((m) => m.id === memberId)
-    if (index !== -1) {
-      // Update the member
-      members.value[index] = {
-        ...members.value[index],
-        ...sanitizedFormData,
-        updatedAt: now,
-      }
-    }
-  } else {
-    // Create new member
-    const newMember: Member = {
-      id: memberId,
-      ...sanitizedFormData,
-      createdAt: now,
-      updatedAt: now,
-    }
-
-    members.value.push(newMember)
-    selectedMember.value = newMember
-
-    // Add to parent union if specified
-    if (newParentUnionId) {
-      const parentUnion = unions.value.find((u) => u.id === newParentUnionId)
-      if (parentUnion) {
-        if (!parentUnion.childrenIds) {
-          parentUnion.childrenIds = []
-        }
-        if (!parentUnion.childrenIds.includes(memberId)) {
-          parentUnion.childrenIds.push(memberId)
-        }
-      }
-    }
-  }
-
-  // Rebuild parentUnionIds to ensure consistency
-  members.value = buildParentUnionIds(members.value, unions.value)
 
   saveToLocalStorage()
   closeModal()
@@ -766,8 +493,7 @@ const closeModal = () => {
 }
 
 const handleFileImportSuccess = (familyData: FamilyData) => {
-  members.value = familyData.members
-  unions.value = familyData.unions
+  setFamilyData(familyData)
   saveToLocalStorage()
 }
 
@@ -780,24 +506,22 @@ const cancelDestroyAll = () => {
 }
 
 const destroyAllMembers = () => {
-  // Clear all reactive state
-  members.value = []
-  unions.value = []
+  // Clear all data using the composable
+  clearAllData()
+
+  // Clear UI state
   selectedMember.value = null
   showAddForm.value = false
   editingMember.value = null
   memberToDelete.value = null
   showDestroyAllModal.value = false
 
-  // Save empty state to localStorage
-  saveToLocalStorage()
-
   // Reset the FamilyTree component
   familyTreeRef.value?.reset()
 }
 
 const searchMembers = (query: string) => {
-  searchResults.value = members.value.filter((member) => {
+  searchResults.value = getAllMembers().filter((member) => {
     return getFullName(member).toLowerCase().includes(query.toLowerCase())
   })
   console.log('Search results:', searchResults.value)
@@ -812,9 +536,9 @@ const closeSettingsModal = () => {
   showSettingsModal.value = false
 }
 
-const saveSettings = (newSettings: any) => {
+const handleSaveSettings = (newSettings: any) => {
   settings.value = { ...settings.value, ...newSettings }
-  localStorage.setItem('family-settings', JSON.stringify(settings.value))
+  saveSettings(settings.value)
   closeSettingsModal()
 }
